+++ src/frontend/src/app/components/financeiro/condicao-pagamento/condicao-pagamento-grid/condicao-pagamento-grid.component.ts
@@ -5,6 +5,8 @@ import { CondicaoPagamentoService } from '../condicao-pagamento.service';
 import { Order, PageRequest } from '../../../base/model/page-request';
 import { CondicaoPagamentoGridDTO } from '../model/condicao-pagamento-grid.dto';
 import { AuthService } from '../../../base/auth/auth-service';
+import { DialogService } from '../../../base/dialog/dialog.service';
+import { DialogResult } from '../../../base/dialog/dialog.model';
 import { TableComponent } from '../../../base/table/table.component';
 import { ColumnModel } from '../../../base/table/column.model';
 import { ActionModel } from '../../../base/table/action.model';
@@ -96,6 +98,7 @@ export class CondicaoPagamentoGridComponent {
 
   private service: CondicaoPagamentoService = inject(CondicaoPagamentoService);
   private auth: AuthService = inject(AuthService);
+  private dialogService: DialogService = inject(DialogService);
 
   constructor() {
     const canView = this.auth.hasAuthorityVisualizarToModulo(
@@ -122,7 +125,16 @@ export class CondicaoPagamentoGridComponent {
         icon: 'delete',
         title: $localize`Excluir`,
         action: (element: CondicaoPagamentoGridDTO) => {
-          this.service.delete(element.id!).subscribe(() => this.listar());
+          this.dialogService
+            .showYesNo(
+              $localize`Confirmar Exclusão`,
+              $localize`Deseja realmente excluir o registro selecionado?`
+            )
+            .subscribe((result) => {
+              if (result === DialogResult.YES) {
+                this.service.delete(element.id!).subscribe(() => this.listar());
+              }
+            });
         },
       });
     }
