+++ src/frontend/src/app/components/financeiro/titulo-categoria/titulo-categoria-grid/titulo-categoria-grid.component.ts
@@ -5,6 +5,8 @@ import { TituloCategoriaService } from '../titulo-categoria.service';
 import { Order, PageRequest } from '../../../base/model/page-request';
 import { TituloCategoriaGridDTO } from '../model/titulo-categoria-grid.dto';
 import { AuthService } from '../../../base/auth/auth-service';
+import { DialogService } from '../../../base/dialog/dialog.service';
+import { DialogResult } from '../../../base/dialog/dialog.model';
 import { TableComponent } from '../../../base/table/table.component';
 import { ColumnModel } from '../../../base/table/column.model';
 import { ActionModel } from '../../../base/table/action.model';
@@ -110,6 +112,7 @@ export class TituloCategoriaGridComponent {
 
   private service: TituloCategoriaService = inject(TituloCategoriaService);
   private auth: AuthService = inject(AuthService);
+  private dialogService: DialogService = inject(DialogService);
 
   constructor() {
     const canView = this.auth.hasAuthorityVisualizarToModulo(
@@ -136,7 +139,16 @@ export class TituloCategoriaGridComponent {
         icon: 'delete',
         title: $localize`Excluir`,
         action: (element: TituloCategoriaGridDTO) => {
-          this.service.delete(element.id!).subscribe(() => this.listar());
+          this.dialogService
+            .showYesNo(
+              $localize`Confirmar Exclusão`,
+              $localize`Deseja realmente excluir o registro selecionado?`
+            )
+            .subscribe((result) => {
+              if (result === DialogResult.YES) {
+                this.service.delete(element.id!).subscribe(() => this.listar());
+              }
+            });
         },
       });
     }
